Split up jobs.py into a package.
